Add AppleIntelPIIXATA and IOATAFamily projects
I bumped the version number on the IOATAFamily project so I can
ensure that the version tagged on GitHub will be used, instead of
the version on opensource.apple.com. (I used the fifth version
element to ensure that it will not conflict with a future release
of IOATAFamily by Apple. Apple generally does not use the fifth
version element for public source releases.)